refactor: split dashboard modal js into modules
