+++ static/js/dashboard_modal.js
@@ -1,3 +1,6 @@
+import { updateAggToggleUI, updateAverageButtonUI, updateMathFieldUI, refreshColumnTags, updateValueResult, populateFieldDropdown, updateColumnOptions, initColumnSelect, initOperationSelect } from "./value_widgets.js";
+import { updateTablePreview } from "./table_widgets.js";
+
 let dashboardTrigger = null;
 let escHandler = (e) => {
   if (e.key === 'Escape') {
@@ -60,15 +63,6 @@ let filteredSearchInputEl, filteredLimitInputEl, filteredRecordsContainer;
 let tableTitleInputEl, tableCreateBtnEl, tablePreviewEl, tablePreviewBodyEl, tablePreviewHeaderEl;
 let tableData = [];
 
-function isNumericField(val) {
-  if (!val) return false;
-  const [table, field] = val.split(':');
-  return (
-    FIELD_SCHEMA[table] &&
-    FIELD_SCHEMA[table][field] &&
-    FIELD_SCHEMA[table][field].type === 'number'
-  );
-}
 
 function getNonTextTypes() {
   const types = new Set();
@@ -78,77 +72,6 @@ function getNonTextTypes() {
   return Array.from(types).filter(t => t !== 'text' && t !== 'textarea' && t !== 'url');
 }
 
-function getLabelField(table) {
-  const schema = FIELD_SCHEMA[table] || {};
-  const fields = Object.keys(schema);
-  if (!fields.length) return null;
-  if (fields.includes(table)) return table;
-  return fields.length > 1 ? fields[1] : fields[0];
-}
-
-function toggleDisabled(label, input, disabled) {
-  if (!label || !input) return;
-  input.disabled = disabled;
-  if (disabled) {
-    label.classList.add('opacity-50', 'pointer-events-none');
-  } else {
-    label.classList.remove('opacity-50', 'pointer-events-none');
-  }
-}
-
-function updateAggToggleUI() {
-  if (aggToggle1El) {
-    const sumInput = aggToggle1El.querySelector('input[value="sum"]');
-    const countInput = aggToggle1El.querySelector('input[value="count"]');
-    const sumLabel = sumInput ? sumInput.parentElement : null;
-    const numeric = isNumericField(mathField1);
-    toggleDisabled(sumLabel, sumInput, !numeric);
-    if (!numeric && sumInput && sumInput.checked && countInput) {
-      countInput.checked = true;
-      agg1 = 'count';
-    }
-  }
-  if (aggToggle2El) {
-    const sumInput = aggToggle2El.querySelector('input[value="sum"]');
-    const countInput = aggToggle2El.querySelector('input[value="count"]');
-    const sumLabel = sumInput ? sumInput.parentElement : null;
-    const numeric = isNumericField(mathField2);
-    toggleDisabled(sumLabel, sumInput, !numeric);
-    if (!numeric && sumInput && sumInput.checked && countInput) {
-      countInput.checked = true;
-      agg2 = 'count';
-    }
-  }
-}
-
-function updateAverageButtonUI() {
-  if (!mathOpContainer) return;
-  const avgInput = mathOpContainer.querySelector('input[value="average"]');
-  const avgLabel = avgInput ? avgInput.parentElement : null;
-  const numeric = isNumericField(mathField1);
-  toggleDisabled(avgLabel, avgInput, !numeric);
-  if (!numeric && avgInput && avgInput.checked) {
-    avgInput.checked = false;
-    mathOperation = null;
-  }
-}
-
-function updateMathFieldUI() {
-  if (!mathField1Container || !mathField2Container) return;
-  if (selectedOperation === 'math') {
-    mathField1Container.classList.remove('hidden');
-    if (mathOperation && mathOperation !== 'average') {
-      mathField2Container.classList.remove('hidden');
-    } else {
-      mathField2Container.classList.add('hidden');
-      mathField2 = null;
-      if (mathSelect2Label) mathSelect2Label.textContent = 'Select Field';
-    }
-  } else {
-    mathField1Container.classList.add('hidden');
-    mathField2Container.classList.add('hidden');
-  }
-}
 
 function updateChartUI() {
   if (!chartTypeEl || !chartXFieldContainer || !chartYFieldContainer || !chartAggContainer || !chartOrientContainer || !chartTitleInputEl || !chartCreateBtnEl) return;
@@ -225,145 +148,6 @@ function updateChartTitle() {
   chartTitleInputEl.value = defaultTitle;
 }
 
-function updateTablePreview() {
-  if (!tablePreviewEl || !tablePreviewBodyEl || !tableTitleInputEl || !tableCreateBtnEl) return;
-  tablePreviewBodyEl.innerHTML = '';
-  tablePreviewEl.classList.add('hidden');
-  tableTitleInputEl.classList.add('hidden');
-  tableCreateBtnEl.classList.add('hidden');
-  if (selectCountToggleBtn) selectCountToggleBtn.classList.add('hidden');
-  if (selectCountOptions) selectCountOptions.classList.add('hidden');
-  if (selectCountFieldContainer) selectCountFieldContainer.classList.add('hidden');
-  if (topFieldToggleBtn) topFieldToggleBtn.classList.add('hidden');
-  if (topFieldOptions) topFieldOptions.classList.add('hidden');
-  if (topFieldContainer) topFieldContainer.classList.add('hidden');
-  if (topDirectionContainer) topDirectionContainer.classList.add('hidden');
-  if (filteredRecordsContainer) filteredRecordsContainer.classList.add('hidden');
-  if (tableType === 'base-count') {
-    if (tablePreviewHeaderEl)
-      tablePreviewHeaderEl.innerHTML = '<th class="px-2 py-1 text-left">Table</th><th class="px-2 py-1 text-left">Count</th>';
-    fetch('/dashboard/base-count')
-      .then(r => r.json())
-      .then(rows => {
-        tableData = rows || [];
-        tableData.forEach(r => {
-          const tr = document.createElement('tr');
-          tr.innerHTML = `<td class="px-2 py-1">${r.table}</td><td class="px-2 py-1">${r.count}</td>`;
-          tablePreviewBodyEl.appendChild(tr);
-        });
-        tablePreviewEl.classList.remove('hidden');
-        tableTitleInputEl.placeholder = 'Base Table Counts';
-        tableTitleInputEl.value = 'Base Table Counts';
-        tableTitleInputEl.classList.remove('hidden');
-        tableCreateBtnEl.classList.remove('hidden');
-      })
-      .catch(() => {
-        tableData = [];
-        tablePreviewBodyEl.innerHTML = '<tr><td colspan="2" class="px-2 py-1">Error</td></tr>';
-        tablePreviewEl.classList.remove('hidden');
-      });
-  } else if (tableType === 'select-count') {
-    if (selectCountToggleBtn) selectCountToggleBtn.classList.remove('hidden');
-    if (!selectCountField) {
-      if (selectCountFieldContainer) selectCountFieldContainer.classList.remove('hidden');
-      return;
-    }
-    if (selectCountFieldContainer) selectCountFieldContainer.classList.remove('hidden');
-    const [tbl, fld] = selectCountField.split(':');
-    if (tablePreviewHeaderEl)
-      tablePreviewHeaderEl.innerHTML = '<th class="px-2 py-1 text-left">Value</th><th class="px-2 py-1 text-left">Count</th>';
-    fetch(`/${tbl}/field-distribution?field=${encodeURIComponent(fld)}`)
-      .then(r => r.json())
-      .then(data => {
-        tableData = Object.entries(data).map(([value, count]) => ({ value, count }));
-        tableData.forEach(r => {
-          const tr = document.createElement('tr');
-          tr.innerHTML = `<td class="px-2 py-1">${r.value}</td><td class="px-2 py-1">${r.count}</td>`;
-          tablePreviewBodyEl.appendChild(tr);
-        });
-      tablePreviewEl.classList.remove('hidden');
-      const fieldLabel = fld;
-      tableTitleInputEl.placeholder = `Value Counts for ${fieldLabel}`;
-      tableTitleInputEl.value = `Value Counts for ${fieldLabel}`;
-      tableTitleInputEl.classList.remove('hidden');
-      tableCreateBtnEl.classList.remove('hidden');
-      })
-      .catch(() => {
-        tableData = [];
-        tablePreviewBodyEl.innerHTML = '<tr><td colspan="2" class="px-2 py-1">Error</td></tr>';
-        tablePreviewEl.classList.remove('hidden');
-      });
-  } else if (tableType === 'top-numeric') {
-    if (topFieldToggleBtn) topFieldToggleBtn.classList.remove('hidden');
-    if (!topNumericField) {
-      if (topFieldContainer) topFieldContainer.classList.remove('hidden');
-      if (topDirectionContainer) topDirectionContainer.classList.remove('hidden');
-      return;
-    }
-    if (topFieldContainer) topFieldContainer.classList.remove('hidden');
-    if (topDirectionContainer) topDirectionContainer.classList.remove('hidden');
-    const [tbl, fld] = topNumericField.split(':');
-    if (tablePreviewHeaderEl)
-      tablePreviewHeaderEl.innerHTML = '<th class="px-2 py-1 text-left">ID</th><th class="px-2 py-1 text-left">Value</th>';
-    fetch(`/dashboard/top-numeric?table=${encodeURIComponent(tbl)}&field=${encodeURIComponent(fld)}&direction=${topDirection}&limit=5`)
-      .then(r => r.json())
-      .then(rows => {
-        tableData = rows || [];
-        tableData.forEach(r => {
-          const tr = document.createElement('tr');
-          tr.innerHTML = `<td class="px-2 py-1"><a href="/${tbl}/${r.id}" class="text-blue-600 underline">${r.id}</a></td><td class="px-2 py-1">${r.value}</td>`;
-          tablePreviewBodyEl.appendChild(tr);
-        });
-      tablePreviewEl.classList.remove('hidden');
-      const prefix = topDirection === 'desc' ? 'Top' : 'Bottom';
-      tableTitleInputEl.placeholder = `${prefix} ${tableData.length} of ${fld}`;
-      tableTitleInputEl.value = `${prefix} ${tableData.length} of ${fld}`;
-      tableTitleInputEl.classList.remove('hidden');
-      tableCreateBtnEl.classList.remove('hidden');
-      })
-      .catch(() => {
-        tableData = [];
-        tablePreviewBodyEl.innerHTML = '<tr><td colspan="2" class="px-2 py-1">Error</td></tr>';
-        tablePreviewEl.classList.remove('hidden');
-      });
-  } else if (tableType === 'filtered-records') {
-    if (!filteredTable) {
-      if (filteredRecordsContainer) filteredRecordsContainer.classList.remove('hidden');
-      return;
-    }
-    if (filteredRecordsContainer) filteredRecordsContainer.classList.remove('hidden');
-    const labelField = getLabelField(filteredTable);
-    if (tablePreviewHeaderEl)
-      tablePreviewHeaderEl.innerHTML = `<th class="px-2 py-1 text-left">ID</th><th class="px-2 py-1 text-left">${labelField || 'Label'}</th>`;
-    const params = new URLSearchParams({
-      table: filteredTable,
-      search: filteredSearch || '',
-      order_by: filteredSort || '',
-      limit: filteredLimit || 10
-    });
-    fetch(`/dashboard/filtered-records?${params.toString()}`)
-      .then(r => r.json())
-      .then(rows => {
-        tableData = rows || [];
-        tableData.forEach(r => {
-          const tr = document.createElement('tr');
-          const label = r[labelField];
-          tr.innerHTML = `<td class="px-2 py-1"><a href="/${filteredTable}/${r.id}" class="text-blue-600 underline">${r.id}</a></td><td class="px-2 py-1">${label ?? ''}</td>`;
-          tablePreviewBodyEl.appendChild(tr);
-        });
-        tablePreviewEl.classList.remove('hidden');
-        tableTitleInputEl.placeholder = `Filtered ${filteredTable}`;
-        tableTitleInputEl.value = `Filtered ${filteredTable}`;
-        tableTitleInputEl.classList.remove('hidden');
-        tableCreateBtnEl.classList.remove('hidden');
-      })
-      .catch(() => {
-        tableData = [];
-        tablePreviewBodyEl.innerHTML = '<tr><td colspan="2" class="px-2 py-1">Error</td></tr>';
-        tablePreviewEl.classList.remove('hidden');
-      });
-  }
-}
 
 function setActiveTab(name) {
   activeTab = name;
@@ -380,90 +164,6 @@ function initDashboardTabs() {
   });
 }
 
-function refreshColumnTags() {
-  if (!columnToggleBtn) return;
-  if (!selectedColumn) {
-    if (columnToggleLabel) columnToggleLabel.textContent = 'Select Field';
-    updateValueResult();
-    return;
-  }
-
-  const [table, field] = selectedColumn.split(':');
-  if (columnToggleLabel) columnToggleLabel.textContent = `${table}: ${field}`;
-
-  updateValueResult();
-}
-
-function updateValueResult() {
-  if (!valueResultEl || !resultRowEl) return;
-  if ((selectedOperation === 'sum' || selectedOperation === 'count') && selectedColumn) {
-    const [table, field] = selectedColumn.split(':');
-    const endpoint = selectedOperation === 'sum' ? 'sum-field' : 'count-nonnull';
-    const key = selectedOperation === 'sum' ? 'sum' : 'count';
-    resultRowEl.classList.remove('hidden');
-    if (titleInputEl) {
-      const defaultTitle = `${selectedOperation === 'sum' ? 'Sum' : 'Count'} of ${field}`;
-      titleInputEl.placeholder = defaultTitle;
-      titleInputEl.value = defaultTitle;
-    }
-    if (createBtnEl) createBtnEl.classList.remove('hidden');
-    valueResultEl.textContent = 'Calculating…';
-    fetch(`/${table}/${endpoint}?field=${encodeURIComponent(field)}`)
-      .then(res => res.json())
-      .then(data => {
-        valueResultEl.textContent = data[key];
-      })
-      .catch(() => {
-        valueResultEl.textContent = 'Error';
-      });
-  } else if (selectedOperation === 'math' && mathField1 && mathOperation && (mathOperation === 'average' || mathField2)) {
-    resultRowEl.classList.remove('hidden');
-    if (createBtnEl) createBtnEl.classList.remove('hidden');
-    const labels = [mathField1, mathOperation !== 'average' ? mathField2 : null].filter(Boolean).map(v => v.split(':')[1]);
-    const defaultTitle = `${mathOperation.charAt(0).toUpperCase() + mathOperation.slice(1)} of ${labels.join(', ')}`;
-    if (titleInputEl) {
-      titleInputEl.placeholder = defaultTitle;
-      titleInputEl.value = defaultTitle;
-    }
-    valueResultEl.textContent = 'Calculating…';
-    if (mathOperation === 'average') {
-      const [table, field] = mathField1.split(':');
-      Promise.all([
-        fetch(`/${table}/sum-field?field=${encodeURIComponent(field)}`).then(r => r.json()).then(d => d.sum || 0).catch(() => 0),
-        fetch(`/${table}/count-nonnull?field=${encodeURIComponent(field)}`).then(r => r.json()).then(d => d.count || 1).catch(() => 1)
-      ])
-      .then(([s, c]) => { valueResultEl.textContent = c ? s / c : 0; })
-      .catch(() => { valueResultEl.textContent = 'Error'; });
-    } else {
-      const fields = [
-        { val: mathField1, mode: agg1 },
-        { val: mathField2, mode: agg2 }
-      ];
-      Promise.all(fields.map(f => {
-        const [table, field] = f.val.split(':');
-        const endpoint = f.mode === 'sum' ? 'sum-field' : 'count-nonnull';
-        const key = f.mode === 'sum' ? 'sum' : 'count';
-        return fetch(`/${table}/${endpoint}?field=${encodeURIComponent(field)}`)
-          .then(res => res.json())
-          .then(d => d[key] || 0)
-          .catch(() => 0);
-      }))
-      .then(nums => {
-        let result = 0;
-        if (mathOperation === 'add') result = nums[0] + nums[1];
-        else if (mathOperation === 'subtract') result = nums[0] - nums[1];
-        else if (mathOperation === 'multiply') result = nums[0] * nums[1];
-        else if (mathOperation === 'divide') result = nums[0] / (nums[1] || 1);
-        valueResultEl.textContent = result;
-      })
-      .catch(() => { valueResultEl.textContent = 'Error'; });
-    }
-  } else {
-    resultRowEl.classList.add('hidden');
-    valueResultEl.textContent = '';
-    if (createBtnEl) createBtnEl.classList.add('hidden');
-  }
-}
 
 function onCreateWidget(event) {
   if (event) event.preventDefault();
@@ -624,141 +324,6 @@ function onCreateWidget(event) {
     });
 }
 
-function populateFieldDropdown(dropdown, restrictNumeric, allowedTypes, callback, excludeTypes = []) {
-  if (!dropdown) return;
-  dropdown.innerHTML = '';
-  const search = document.createElement('input');
-  search.type = 'text';
-  search.placeholder = 'Search...';
-  search.className = 'w-full px-2 py-1 border rounded text-sm mb-2';
-  search.addEventListener('input', function() {
-    const v = this.value.toLowerCase();
-    [...dropdown.querySelectorAll('label')].forEach(l => l.classList.toggle('hidden', !l.textContent.toLowerCase().includes(v)));
-  });
-  dropdown.appendChild(search);
-
-  Object.keys(FIELD_SCHEMA).forEach(table => {
-    const tableSchema = FIELD_SCHEMA[table] || {};
-    const fields = Object.keys(tableSchema);
-    fields.forEach(field => {
-      const type = tableSchema[field] ? tableSchema[field].type : '';
-      if (restrictNumeric && type !== 'number') return;
-      if (allowedTypes && !allowedTypes.includes(type)) return;
-      if (excludeTypes && excludeTypes.includes(type)) return;
-      const val = `${table}:${field}`;
-      const label = document.createElement('label');
-      label.className = 'flex items-center space-x-2';
-      const input = document.createElement('input');
-      input.type = 'radio';
-      input.name = 'fieldSelect';
-      input.value = val;
-      input.className = 'rounded border-gray-300 text-blue-600 shadow-sm focus:ring-blue-500';
-      input.addEventListener('change', () => {
-        callback(val);
-        dropdown.classList.add('hidden');
-      });
-      const span = document.createElement('span');
-      span.className = 'text-sm';
-      span.innerHTML = `<strong>${table}</strong>: ${field} <span class="text-blue-600 text-xs">(${type})</span>`;
-      label.appendChild(input);
-      label.appendChild(span);
-      dropdown.appendChild(label);
-    });
-  });
-}
-
-function updateColumnOptions() {
-  if (!columnToggleBtn || !columnDropdown) return;
-
-  // hide everything first
-  columnToggleBtn.classList.add('hidden');
-  columnDropdown.classList.add('hidden');
-  if (mathSelect1Btn) mathSelect1Btn.parentElement.parentElement.classList.add('hidden');
-  if (mathSelect2Btn) mathSelect2Btn.parentElement.parentElement.classList.add('hidden');
-  if (mathOpContainer) mathOpContainer.classList.add('hidden');
-
-  if (!selectedOperation) {
-    selectedColumn = null;
-    mathField1 = null;
-    mathField2 = null;
-    mathOperation = null;
-    updateAggToggleUI();
-    updateAverageButtonUI();
-    updateValueResult();
-    return;
-  }
-
-  if (selectedOperation === 'sum' || selectedOperation === 'count') {
-    columnToggleBtn.classList.remove('hidden');
-    populateFieldDropdown(columnDropdown, selectedOperation === 'sum', null, val => {
-      selectedColumn = val;
-      refreshColumnTags();
-    });
-    refreshColumnTags();
-  } else if (selectedOperation === 'math') {
-    if (mathSelect1Btn) mathSelect1Btn.parentElement.parentElement.classList.remove('hidden');
-    if (mathOpContainer) mathOpContainer.classList.remove('hidden');
-    populateFieldDropdown(mathSelect1Options, false, null, val => {
-      mathField1 = val;
-      if (mathSelect1Label) {
-        const [t,f] = val.split(':');
-        mathSelect1Label.textContent = `${t}: ${f}`;
-      }
-      updateAggToggleUI();
-      updateAverageButtonUI();
-      updateValueResult();
-    });
-    populateFieldDropdown(mathSelect2Options, false, null, val => {
-      mathField2 = val;
-      if (mathSelect2Label) {
-        const [t,f] = val.split(':');
-        mathSelect2Label.textContent = `${t}: ${f}`;
-      }
-      updateAggToggleUI();
-      updateValueResult();
-    });
-  }
-
-  updateAggToggleUI();
-  updateAverageButtonUI();
-  updateMathFieldUI();
-  updateValueResult();
-}
-
-function initColumnSelect() {
-  columnToggleBtn = document.getElementById('columnSelectDashboardToggle');
-  columnToggleLabel = columnToggleBtn ? columnToggleBtn.querySelector('.selected-label') : null;
-  columnDropdown = document.getElementById('columnSelectDashboardOptions');
-  if (!columnToggleBtn || !columnDropdown) return;
-
-  columnToggleBtn.addEventListener('click', e => {
-    e.stopPropagation();
-    columnDropdown.classList.toggle('hidden');
-  });
-  document.addEventListener('click', e => {
-    if (!columnDropdown.contains(e.target) && e.target !== columnToggleBtn) {
-      columnDropdown.classList.add('hidden');
-    }
-  });
-  columnDropdown.addEventListener('click', e => e.stopPropagation());
-
-  updateColumnOptions();
-}
-
-function initOperationSelect() {
-  const opSelect = document.getElementById('dashboardOperation');
-  if (!opSelect) return;
-  opSelect.addEventListener('change', () => {
-    const checked = opSelect.querySelector('input[name="dashboardOperation"]:checked');
-    selectedOperation = checked ? checked.value : null;
-    selectedColumn = null;
-    mathField1 = null;
-    mathField2 = null;
-    mathOperation = null;
-    updateColumnOptions();
-    updateMathFieldUI();
-  });
-}
 
 function initDashboardModal() {
   initDashboardTabs();
